requirements.txt
# Core data processing
pandas>=2.3.1
numpy>=1.26.0

# Text processing and embeddings
langchain>=0.3.26
langchain-community>=0.0.10
langchain-huggingface>=0.0.6
sentence-transformers>=5.0.0
faiss-cpu>=1.11.0
torch>=2.7.1
transformers>=4.53.1
accelerate>=1.8.1

# LLM and RAG
huggingface-hub[hf_xet]>=0.20.0
tabulate>=0.9.0
python-dotenv>=1.0.0

# Utilities
tqdm>=4.67.0

# Development
black>=23.7.0
flake8>=6.1.0
pytest>=7.4.0